Split shared TypeScript into contracts and dashboard-runtime
Add @helpudoc/contracts and @helpudoc/dashboard-runtime; keep @helpudoc/shared
as relative re-exports. Wire frontend/backend/Docker/compose and CI path filters.
