+++ backend/Dockerfile.gke
@@ -3,6 +3,9 @@ FROM node:20-bookworm-slim AS base
 
 WORKDIR /app
 
+# Local `file:` dependency (@helpudoc/contracts) resolves to ../packages/contracts from this package.json directory.
+COPY packages/contracts /packages/contracts
+
 # Copy dependency manifests first for better layer caching
 COPY backend/package*.json ./
 COPY backend/tsconfig.json ./
