+++ backend/package.json
@@ -17,6 +17,7 @@
   "description": "",
   "dependencies": {
     "@aws-sdk/client-s3": "^3.914.0",
+    "@helpudoc/contracts": "file:../packages/contracts",
     "@hocuspocus/extension-database": "^3.4.3",
     "@hocuspocus/server": "^3.4.3",
     "@types/express-session": "^1.18.2",
