+++ frontend/src/components/dashboard/DashboardCanvas.tsx
@@ -10,7 +10,7 @@ import {
   type DashboardRow,
   type DatasetSchemaColumn,
   type FilterValues,
-} from '@helpudoc/shared/dashboard';
+} from '@helpudoc/dashboard-runtime';
 import DashboardFilters from './DashboardFilters';
 
 function normalizePath(value: string): string {
