+++ packages/contracts/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@helpudoc/contracts",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./types": "./src/types.ts",
+    "./agentStream": "./src/agentStream.ts"
+  }
+}
