+++ packages/contracts/src/types.ts
@@ -1,0 +1,336 @@
+export interface Workspace {
+  id: string;
+  name: string;
+  lastUsed: string;
+  slug?: string;
+  role?: 'owner' | 'editor' | 'viewer';
+  canEdit?: boolean;
+  skipPlanApprovals?: boolean;
+  createdAt?: string;
+  updatedAt?: string;
+}
+
+export interface File {
+  id: string;
+  name: string;
+  workspaceId?: string;
+  storageType?: 'local' | 's3';
+  path?: string;
+  mimeType?: string | null;
+  publicUrl?: string | null;
+  content?: string;
+  understandingStatus?: DerivedArtifactStatus | null;
+  understandingMode?: DerivedArtifactMode | null;
+  understandingError?: string | null;
+  derivedArtifactFileId?: number | null;
+}
+
+export type GoogleDrivePickerScope = 'recent' | 'my-drive' | 'shared';
+
+export type GoogleDriveIconHint = 'docs' | 'sheets' | 'slides' | 'pdf' | 'image' | 'file';
+
+export interface GoogleDrivePickerItem {
+  id: string;
+  name: string;
+  mimeType: string;
+  webViewUrl?: string | null;
+  modifiedTime?: string | null;
+  ownerNames?: string[];
+  size?: string | null;
+  iconHint: GoogleDriveIconHint;
+  scope?: GoogleDrivePickerScope;
+}
+
+export interface GoogleDriveSearchResult {
+  files: GoogleDrivePickerItem[];
+  nextPageToken?: string | null;
+}
+
+export type DerivedArtifactStatus = 'pending' | 'partial' | 'ready' | 'failed' | 'superseded';
+export type DerivedArtifactMode = 'part' | 'parser' | 'hybrid';
+export type AttachmentPrepStatus = 'pending' | 'running' | 'ready' | 'failed';
+
+export interface FileContextRef {
+  sourceFileId: number;
+  sourceName: string;
+  sourceMimeType?: string | null;
+  sourceVersionFingerprint: string;
+  artifactId: string;
+  artifactVersion: number;
+  derivedArtifactFileId?: number | null;
+  derivedArtifactPath?: string | null;
+  effectiveMode: DerivedArtifactMode;
+  status: DerivedArtifactStatus;
+  summary?: string | null;
+  lastError?: string | null;
+}
+
+export interface AgentPersona {
+  name: string;
+  displayName: string;
+  description?: string;
+}
+
+export interface ConversationSummary {
+  id: string;
+  workspaceId: string;
+  persona: string;
+  title: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface ToolEvent {
+  id: string;
+  name: string;
+  status: 'running' | 'completed' | 'error';
+  summary?: string;
+  startedAt: string;
+  finishedAt?: string;
+  outputFiles?: ToolOutputFile[];
+  relatedFiles?: ToolOutputFile[];
+}
+
+export interface ToolOutputFile {
+  path: string;
+  mimeType?: string | null;
+  size?: number;
+}
+
+/** Stream event + durable dashboard package; no TTL or live URL. */
+export interface DashboardArtifactInfo {
+  dashboardPath: string;
+  workspaceId?: string;
+  dashboardId?: string;
+  title?: string;
+  status: 'generating' | 'ready' | 'error';
+}
+
+export interface InterruptChoice {
+  id: string;
+  label: string;
+  description?: string;
+  value: string;
+}
+
+export interface InterruptQuestionOption {
+  id: string;
+  label: string;
+  description?: string;
+  value: string;
+}
+
+export interface InterruptQuestion {
+  id: string;
+  header: string;
+  question: string;
+  options?: InterruptQuestionOption[];
+}
+
+export type InterruptAnswerValue = string | string[];
+export type InterruptAnswersByQuestionId = Record<string, InterruptAnswerValue>;
+
+export interface InterruptAction {
+  id: string;
+  label: string;
+  style?: 'primary' | 'secondary' | 'danger';
+  inputMode?: 'none' | 'text';
+  placeholder?: string;
+  submitLabel?: string;
+  confirm?: boolean;
+  value?: string;
+  payload?: Record<string, unknown>;
+}
+
+export interface InterruptResponseSpec {
+  inputMode?: 'none' | 'text' | 'choice' | 'text_or_choice';
+  multiple?: boolean;
+  submitLabel?: string;
+  placeholder?: string;
+  allowDismiss?: boolean;
+  dismissLabel?: string;
+  choices?: InterruptChoice[];
+  questions?: InterruptQuestion[];
+}
+
+export interface PendingInterrupt {
+  kind?: 'approval' | 'clarification';
+  interruptId?: string;
+  title?: string;
+  description?: string;
+  stepIndex?: number;
+  stepCount?: number;
+  actions?: InterruptAction[];
+  actionRequests?: Array<{ name?: string; args?: Record<string, unknown> }>;
+  reviewConfigs?: Array<{ action_name?: string; allowed_decisions?: string[] }>;
+  responseSpec?: InterruptResponseSpec;
+  displayPayload?: Record<string, unknown>;
+}
+
+export interface ConversationMessageMetadata {
+  thinkingText?: string;
+  toolEvents?: ToolEvent[];
+  bodySource?: 'assistant' | 'summary';
+  runId?: string;
+  status?: 'queued' | 'running' | 'awaiting_approval' | 'completed' | 'failed' | 'cancelled';
+  attachmentJobId?: string;
+  attachmentPrepStatus?: AttachmentPrepStatus;
+  attachmentPrepError?: string;
+  pendingInterrupt?: PendingInterrupt;
+  runPolicy?: {
+    skill?: string;
+    requiresHitlPlan?: boolean;
+    requiresArtifacts?: boolean;
+    requiredArtifactsMode?: string;
+    prePlanSearchLimit?: number;
+    prePlanSearchUsed?: number;
+  };
+  fileContextRefs?: FileContextRef[];
+}
+
+export interface ConversationMessage {
+  id: number | string;
+  conversationId: string;
+  sender: 'user' | 'agent';
+  text: string;
+  createdAt: string;
+  updatedAt?: string;
+  turnId?: string;
+  thinkingText?: string;
+  toolEvents?: ToolEvent[];
+  metadata?: ConversationMessageMetadata | null;
+}
+
+export interface SkillDefinition {
+  id: string;
+  name: string;
+  description?: string;
+  valid: boolean;
+  error?: string;
+  warning?: string;
+}
+
+export interface ReflectionScorecard {
+  outcome: number;
+  reliability: number;
+  friction: number;
+}
+
+export interface ReflectionRecommendation {
+  id: string;
+  title: string;
+  detail: string;
+  priority: 'high' | 'medium' | 'low';
+}
+
+export interface ReflectionConversationSample {
+  conversationId: string;
+  workspaceId?: string | null;
+  workspaceName?: string | null;
+  userId?: string | null;
+  userDisplayName?: string | null;
+  title?: string | null;
+  status: 'completed' | 'failed' | 'cancelled' | 'awaiting_approval' | 'running' | 'queued';
+  excerpt?: string | null;
+}
+
+export interface ReflectionBreakdown {
+  id: number;
+  reflectionId: number;
+  dimension: 'skill' | 'tool' | 'user' | 'workspace';
+  entityKey: string;
+  label: string;
+  rank: number;
+  metrics: Record<string, unknown>;
+  summary?: string | null;
+}
+
+export interface DailyReflection {
+  id: number;
+  reflectionDate: string;
+  timezone: string;
+  status: 'ready' | 'running' | 'failed';
+  scorecard: ReflectionScorecard;
+  summaryMarkdown: string;
+  metrics: Record<string, unknown>;
+  recommendations: ReflectionRecommendation[];
+  sampledConversations: ReflectionConversationSample[];
+  breakdowns: ReflectionBreakdown[];
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface ReflectionTrendPoint {
+  reflectionDate: string;
+  timezone: string;
+  scorecard: ReflectionScorecard;
+  metrics: Record<string, unknown>;
+}
+
+export type UserMemoryScope = 'global' | 'workspace';
+export type UserMemorySection = 'preferences' | 'context' | 'skill-routing';
+export type UserMemorySuggestionStatus = 'pending' | 'accepted' | 'rejected' | 'stale';
+
+export interface UserMemoryView {
+  globalPreferences: string;
+  globalContext: string;
+  globalSkillRouting: string;
+  workspacePreferences: string;
+  workspaceContext: string;
+  workspaceSkillRouting: string;
+}
+
+export interface UserMemorySuggestion {
+  id: string;
+  userId: string;
+  workspaceId?: string | null;
+  sourceConversationId?: string | null;
+  sourceRunId?: string | null;
+  targetPath: string;
+  targetScope: UserMemoryScope;
+  targetSection: UserMemorySection;
+  baseContentHash: string;
+  proposedContent: string;
+  rationale: string;
+  status: UserMemorySuggestionStatus;
+  reviewedContent?: string | null;
+  reviewedAt?: string | null;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export type SkillEvolutionTargetKind = 'memory_skill_routing' | 'skill_learnings';
+
+export type SkillEvolutionSuggestionStatus = 'pending' | 'accepted' | 'rejected' | 'stale';
+
+export interface SkillEvolutionEvidence {
+  sourceRunIds: string[];
+  sourceConversationIds: string[];
+  workspaceId?: string | null;
+  userId?: string | null;
+  persona?: string | null;
+  skillId?: string | null;
+  transcriptExcerpt?: string | null;
+  telemetrySummary?: string | null;
+}
+
+export interface SkillEvolutionSuggestion {
+  id: string;
+  targetKind: SkillEvolutionTargetKind;
+  memoryUserId: string;
+  memoryTargetPath?: string | null;
+  targetSkillId?: string | null;
+  workspaceId?: string | null;
+  evidence: SkillEvolutionEvidence;
+  rationale: string;
+  baseContentHash: string;
+  /** Snapshot of target file content when the suggestion was created (for admin diff review). */
+  baseContentSnapshot?: string | null;
+  proposedContent: string;
+  status: SkillEvolutionSuggestionStatus;
+  reviewedContent?: string | null;
+  reviewedAt?: string | null;
+  reviewedByUserId?: string | null;
+  createdAt: string;
+  updatedAt: string;
+}
