+++ packages/dashboard-runtime/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@helpudoc/dashboard-runtime",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
